--- src/components/page/header/Notifications.jsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-
-export default function Notifications() {
-  return (
-    <div>
-      お知らせ
-    </div>
-  );
-}
