--- src/components/page/index.js
@@ -1,3 +1,0 @@
-//エンポリポイント
-export {default as categoryList} from './categoryList' 
-export {default as category} from './category'
