--- src/components/page/setting/Notice1.jsx
@@ -1,49 +1,0 @@
-import React from 'react';
-import Button from '@mui/material/Button';
-import Dialog from '@mui/material/Dialog';
-import DialogActions from '@mui/material/DialogActions';
-import DialogContent from '@mui/material/DialogContent';
-import DialogContentText from '@mui/material/DialogContentText';
-import DialogTitle from '@mui/material/DialogTitle';
-
-export default function Notice1() {
-  const [open, setOpen] = React.useState(false);
-
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-   <div>
-      <Button variant="outlined" onClick={handleClickOpen}>
-        2022-01-11 DEKIRUが開設しました。 <br/>卒制企画書完成
-      </Button><Button variant="outlined" onClick={handleClickOpen}>
-        2022-01-11 DEKIRUが開設しました。 <br/>まだまだ課題が沢山あります。
-      </Button>
-      <Dialog
-        open={open}
-        onClose={handleClose}
-        aria-labelledby="alert-dialog-title"
-        aria-describedby="alert-dialog-description"
-      >
-        <DialogTitle id="alert-dialog-title">
-          {"2022-01-11 DEKIRUが開設しました。"}
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText id="alert-dialog-description">
-            本日からDEKIRUを私用することができます。
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose} autoFocus>
-            閉じる
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </div>
-  );
-}
