--- src/components/page/setting/UserSkill.jsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-export default function UserSkill() {
-    return(
-<>レビュ</>
-    );
-}
